refactor: split bench report comparison helpers
